+++ far/fileedit.cpp
@@ -94,7 +94,7 @@ BOOL __stdcall EnumCodePages (const wchar_t *lpwszCodePage)
 
 		Dialog::SendDlgMessage (g_hDlg, DM_LISTSETDATA, g_nID, (LONG_PTR)&data);
 
-		if ( g_nCodepage == dwCP )
+		if ( (unsigned int)g_nCodepage == dwCP )
 		{
 			FarListPos pos;
 
@@ -925,7 +925,7 @@ int FileEditor::ReProcessKey(int Key,int CalledFromControl)
 {
   DWORD FNAttr;
 
-  if (Flags.Check(FFILEEDIT_REDRAWTITLE) && ((Key & 0x00ffffff) < KEY_END_FKEY))
+  if (Flags.Check(FFILEEDIT_REDRAWTITLE) && (((unsigned int)Key & 0x00ffffff) < KEY_END_FKEY))
     ShowConsoleTitle();
 
   // BugZ#488 - Shift=enter
@@ -940,7 +940,7 @@ int FileEditor::ReProcessKey(int Key,int CalledFromControl)
      никак не соответствует обрабатываемой клавише, возникают разномастные
      глюки
   */
-  if((Key >= KEY_MACRO_BASE && Key <= KEY_MACRO_ENDBASE) || (Key>=KEY_OP_BASE && Key <=KEY_OP_ENDBASE)) // исключаем MACRO
+  if(((unsigned int)Key >= KEY_MACRO_BASE && (unsigned int)Key <= KEY_MACRO_ENDBASE) || ((unsigned int)Key>=KEY_OP_BASE && (unsigned int)Key <=KEY_OP_ENDBASE)) // исключаем MACRO
   {
     ; //
   }
@@ -1447,7 +1447,7 @@ int FileEditor::ProcessQuitKey(int FirstSave,BOOL NeedQuestion)
   }
 
   FarChDir(strOldCurDir);
-  return GetExitCode() == XC_QUIT;
+  return (unsigned int)GetExitCode() == XC_QUIT;
 }
